# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuthSys API",
    "version": "23.0.0",
    "description": "Private OAuth 2.0 and OpenID Connect provider. Authorization Code + PKCE is the supported browser flow."
  },
  "servers": [
    { "url": "https://testmono-user.duckdns.org", "description": "Current HTTPS deployment" },
    { "url": "http://localhost:8080", "description": "Local Docker fallback" }
  ],
...
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List active sessions",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Sanitized session list" } }
      }
    },
    "/api/sessions/{sessionId}": {
      "delete": {
        "tags": ["Sessions"],
        "summary": "Revoke one owned session",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [{ "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } }],
        "responses": {
          "200": { "description": "Session revoked" },
          "404": { "description": "Owned session not found" }
        }
      }
    },
...
    "/api/users/export": {
      "get": {
        "tags": ["Privacy"],
        "summary": "Export the current user's retained data",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "PDPA data export" } }
      }
    },
    "/api/users/account": {
      "delete": {
        "tags": ["Privacy"],
        "summary": "Deactivate and anonymize the current account",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Account anonymized and credentials/grants revoked" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "connect.sid" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "clientBasic": { "type": "http", "scheme": "basic", "description": "OAuth client_id and client_secret" }
    },
...
    "responses": {
...
      "Unauthorized": { "description": "Authentication or token validation failed" },
...
    }
  }
}